Refactor Person Search By Last/Full Name Spec Files;
