+++ spec/services/person_search_by_name_query_builder_part_one_spec.rb
@@ -13,8 +13,6 @@ describe PersonSearchByNameQueryBuilderPartOne do
     let(:full_name_without_suffix_query) do
       PersonSearchResultBuilder.new.fs_full_name_without_suffix_query_part_one
     end
-    let(:last_name_query) { PersonSearchResultBuilder.new.fs_last_name_query_part_one }
-    let(:first_name_query) { PersonSearchResultBuilder.new.fs_first_name_query_part_one }
 
     let(:full_name_params) do
       {
@@ -33,20 +31,6 @@ describe PersonSearchByNameQueryBuilderPartOne do
       }
     end
 
-    let(:last_name_params) do
-      {
-        last_name: last_name,
-        suffix: suffix
-      }
-    end
-
-    let(:first_name_params) do
-      {
-        first_name: first_name,
-        suffix: suffix
-      }
-    end
-
     context 'returns hash' do
       it 'with full name query' do
         query_builder = QueryBuilder.new(person_search_fields: full_name_params)
@@ -60,22 +44,6 @@ describe PersonSearchByNameQueryBuilderPartOne do
         query_builder = QueryBuilder.new(person_search_fields: full_name_without_suffix_params)
         query = query_builder.extend(described_class).query
         expect(query.as_json).to eq full_name_without_suffix_query['query']
-      end
-    end
-
-    context 'returns hash' do
-      it 'with last name query' do
-        query_builder = QueryBuilder.new(person_search_fields: last_name_params)
-        query = query_builder.extend(described_class).query
-        expect(query.as_json).to eq last_name_query['query']
-      end
-    end
-
-    context 'returns hash' do
-      it 'with first name query' do
-        query_builder = QueryBuilder.new(person_search_fields: first_name_params)
-        query = query_builder.extend(described_class).query
-        expect(query.as_json).to eq first_name_query['query']
       end
     end
   end
