+++ spec/services/person_search_by_name_query_builder_part_two_spec.rb
@@ -9,8 +9,6 @@ describe PersonSearchByNameQueryBuilderPartTwo do
     let(:middle_name) { 'middle name' }
 
     let(:full_name_query) { PersonSearchResultBuilder.new.fs_full_name_query_part_two }
-    let(:last_name_query) { PersonSearchResultBuilder.new.fs_last_name_query_part_two }
-    let(:first_name_query) { PersonSearchResultBuilder.new.fs_first_name_query_part_two }
 
     let(:full_name_params) do
       {
@@ -20,35 +18,11 @@ describe PersonSearchByNameQueryBuilderPartTwo do
       }
     end
 
-    let(:last_name_params) do
-      { last_name: last_name }
-    end
-
-    let(:first_name_params) do
-      { first_name: first_name }
-    end
-
     context 'returns hash' do
       it 'with full name query' do
         query_builder = QueryBuilder.new(person_search_fields: full_name_params)
         query = query_builder.extend(described_class).query
         expect(query.as_json).to eq full_name_query['query']
-      end
-    end
-
-    context 'returns hash' do
-      it 'with last name query' do
-        query_builder = QueryBuilder.new(person_search_fields: last_name_params)
-        query = query_builder.extend(described_class).query
-        expect(query.as_json).to eq last_name_query['query']
-      end
-    end
-
-    context 'returns hash' do
-      it 'with first name query' do
-        query_builder = QueryBuilder.new(person_search_fields: first_name_params)
-        query = query_builder.extend(described_class).query
-        expect(query.as_json).to eq first_name_query['query']
       end
     end
   end
