+++ spec/support/helpers/person_search_by_name_query_builder_helper.rb
@@ -319,1324 +319,6 @@ module PersonSearchByNameQueryBuilderHelper
     ]
   end
 
-  def last_name_functions
-    [
-      {
-        "filter": {
-          "bool": {
-            "must": [
-              {
-                "match": {
-                  "last_name": {
-                    "query": 'last name',
-                    "_name": '1_exact'
-                  }
-                }
-              }
-            ]
-          }
-        },
-        "weight": 16_384
-      },
-      {
-        "filter": {
-          "bool": {
-            "must": [
-              {
-                "match": {
-                  "akas.last_name": {
-                    "query": 'last name',
-                    "_name": '2_aka'
-                  }
-                }
-              }
-            ]
-          }
-        },
-        "weight": 8192
-      },
-      {
-        "filter": {
-          "bool": {
-            "must": [
-              {
-                "match": {
-                  "last_name.phonetic": {
-                    "query": 'last name',
-                    "_name": '3_phonetic'
-                  }
-                }
-              }
-            ]
-          }
-        },
-        "weight": 4096
-      },
-      {
-        "filter": {
-          "bool": {
-            "must": [
-              {
-                "match": {
-                  "last_name_ngram": {
-                    "query": 'last name',
-                    "minimum_should_match": '15%',
-                    "_name": '4_partial'
-                  }
-                }
-              }
-            ]
-          }
-        },
-        "weight": 2048
-      },
-      {
-        "filter": {
-          "bool": {
-            "must": [
-              {
-                "fuzzy": {
-                  "last_name": {
-                    "value": 'last name',
-                    "fuzziness": '5',
-                    "prefix_length": '1',
-                    "max_expansions": '50',
-                    "_name": '5_fuzzy'
-                  }
-                }
-              }
-            ]
-          }
-        },
-        "weight": 1024
-      }
-    ]
-  end
-
-  def first_name_functions
-    [
-      {
-        "filter": {
-          "bool": {
-            "must": [
-              {
-                "match": {
-                  "first_name": {
-                    "query": 'first name',
-                    "_name": '1_m_fst'
-                  }
-                }
-              }
-            ]
-          }
-        },
-        "weight": 16_384
-      },
-      {
-        "filter": {
-          "bool": {
-            "must": [
-              {
-                "multi_match": {
-                  "query": 'suffix',
-                  "operator": 'and',
-                  "fields": %w[
-                    last_name
-                    suffix
-                  ],
-                  "type": 'cross_fields',
-                  "_name": '2_mlt_last_suffix'
-                }
-              }
-            ]
-          }
-        },
-        "weight": 8192
-      },
-      {
-        "filter": {
-          "multi_match": {
-            "query": 'first name',
-            "operator": 'and',
-            "fields": [
-              'akas.first_name',
-              'akas.last_name'
-            ],
-            "type": 'cross_fields',
-            "_name": '3_mlt_aka'
-          }
-        },
-        "weight": 4096
-      },
-      {
-        "filter": {
-          "bool": {
-            "must": [
-              {
-                "match": {
-                  "first_name.diminutive": {
-                    "query": 'first name',
-                    "_name": '4_dim_fst'
-                  }
-                }
-              }
-            ]
-          }
-        },
-        "weight": 2048
-      },
-      {
-        "filter": {
-          "bool": {
-            "must": [
-              {
-                "match": {
-                  "first_name.phonetic": {
-                    "query": 'first name',
-                    "_name": '5_pho_fst'
-                  }
-                }
-              }
-            ]
-          }
-        },
-        "weight": 1024
-      },
-      {
-        "filter": {
-          "bool": {
-            "must": [
-              {
-                "fuzzy": {
-                  "first_name": {
-                    "value": 'first name',
-                    "fuzziness": '5',
-                    "prefix_length": '1',
-                    "max_expansions": '25',
-                    "_name": '6_fz_fst'
-                  }
-                }
-              }
-            ]
-          }
-        },
-        "weight": 512
-      },
-      {
-        "filter": {
-          "bool": {
-            "must": [
-              {
-                "match": {
-                  "first_name_ngram": {
-                    "query": 'first name',
-                    "minimum_should_match": '25%',
-                    "_name": '7_prt_fst'
-                  }
-                }
-              }
-            ]
-          }
-        },
-        "weight": 256
-      },
-      {
-        "filter": {
-          "multi_match": {
-            "query": 'first name',
-            "operator": 'and',
-            "fields": %w[
-              first_name
-              last_name
-            ],
-            "fuzziness": '2',
-            "_name": '7_mlt_fz'
-          }
-        },
-        "weight": 200
-      },
-      {
-        "filter": {
-          "bool": {
-            "must": [
-              {
-                "match": {
-                  "last_name": {
-                    "query": 'first name',
-                    "_name": '9a_rev_lst'
-                  }
-                }
-              }
-            ]
-          }
-        },
-        "weight": 128
-      },
-      {
-        "filter": {
-          "bool": {
-            "must": [
-              {
-                "match": {
-                  "last_name": {
-                    "query": 'first name',
-                    "_name": '9b_rev_prt_lst'
-                  }
-                }
-              }
-            ]
-          }
-        },
-        "weight": 64
-      },
-      {
-        "filter": {
-          "bool": {
-            "must": [
-              {
-                "match": {
-                  "last_name_ngram": {
-                    "query": 'first name',
-                    "minimum_should_match": '25%',
-                    "_name": '11_rev_prt_lst'
-                  }
-                }
-              }
-            ]
-          }
-        },
-        "weight": 16
-      },
-      {
-        "filter": {
-          "match": {
-            "first_name": {
-              "query": 'first name',
-              "_name": 'xb_m_fst'
-            }
-          }
-        },
-        "weight": 8
-      }
-    ]
-  end
-
-  def full_name_functions_part_one
-    full_name_without_suffix_functions_part_one.insert(1, suffix_sub_query)
-  end
-
-  def full_name_without_suffix_functions_part_one
-    [
-      {
-        "filter": {
-          "bool": {
-            "must": [
-              {
-                "match": {
-                  "last_name": {
-                    "query": 'last name',
-                    "_name": '1_m_lst'
-                  }
-                }
-              },
-              {
-                "match": {
-                  "first_name": {
-                    "query": 'first name',
-                    "_name": '1_m_fst'
-                  }
-                }
-              }
-            ]
-          }
-        },
-        "weight": 16_384
-      },
-      {
-        "filter": {
-          "multi_match": {
-            "query": 'last name first name',
-            "operator": 'and',
-            "fields": [
-              'akas.first_name',
-              'akas.last_name'
-            ],
-            "type": 'cross_fields',
-            "_name": '3_mlt_aka'
-          }
-        },
-        "weight": 4096
-      },
-      {
-        "filter": {
-          "bool": {
-            "must": [
-              {
-                "match": {
-                  "last_name": {
-                    "query": 'last name',
-                    "_name": '4_dim_lst'
-                  }
-                }
-              },
-              {
-                "match": {
-                  "first_name.diminutive": {
-                    "query": 'first name',
-                    "_name": '4_dim_fst'
-                  }
-                }
-              }
-            ]
-          }
-        },
-        "weight": 2048
-      },
-      {
-        "filter": {
-          "bool": {
-            "must": [
-              {
-                "match": {
-                  "last_name": {
-                    "query": 'last name',
-                    "_name": '5_pho_lst'
-                  }
-                }
-              },
-              {
-                "match": {
-                  "first_name.phonetic": {
-                    "query": 'first name',
-                    "_name": '5_pho_fst'
-                  }
-                }
-              }
-            ]
-          }
-        },
-        "weight": 1024
-      },
-      {
-        "filter": {
-          "bool": {
-            "must": [
-              {
-                "match": {
-                  "last_name": {
-                    "query": 'last name',
-                    "_name": '6_fz_lst'
-                  }
-                }
-              },
-              {
-                "fuzzy": {
-                  "first_name": {
-                    "value": 'first name',
-                    "fuzziness": '5',
-                    "prefix_length": '1',
-                    "max_expansions": '25',
-                    "_name": '6_fz_fst'
-                  }
-                }
-              }
-            ]
-          }
-        },
-        "weight": 512
-      },
-      {
-        "filter": {
-          "bool": {
-            "must": [
-              {
-                "match": {
-                  "last_name": {
-                    "query": 'last name',
-                    "_name": '7_prt_lst'
-                  }
-                }
-              },
-              {
-                "match": {
-                  "first_name_ngram": {
-                    "query": 'first name',
-                    "minimum_should_match": '25%',
-                    "_name": '7_prt_fst'
-                  }
-                }
-              }
-            ]
-          }
-        },
-        "weight": 256
-      },
-      {
-        "filter": {
-          "multi_match": {
-            "query": 'last name first name',
-            "operator": 'and',
-            "fields": %w[
-              first_name
-              last_name
-            ],
-            "fuzziness": '2',
-            "_name": '7_mlt_fz'
-          }
-        },
-        "weight": 200
-      }
-    ]
-  end
-
-  def last_name_functions_part_one
-    [
-      {
-        "filter": {
-          "bool": {
-            "must": [
-              {
-                "match": {
-                  "last_name": {
-                    "query": 'last name',
-                    "_name": '1_m_lst'
-                  }
-                }
-              }
-            ]
-          }
-        },
-        "weight": 16_384
-      },
-      {
-        "filter": {
-          "bool": {
-            "must": [
-              {
-                "multi_match": {
-                  "query": 'last name suffix',
-                  "operator": 'and',
-                  "fields": %w[
-                    last_name
-                    suffix
-                  ],
-                  "type": 'cross_fields',
-                  "_name": '2_mlt_last_suffix'
-                }
-              }
-            ]
-          }
-        },
-        "weight": 8192
-      },
-      {
-        "filter": {
-          "multi_match": {
-            "query": 'last name',
-            "operator": 'and',
-            "fields": [
-              'akas.first_name',
-              'akas.last_name'
-            ],
-            "type": 'cross_fields',
-            "_name": '3_mlt_aka'
-          }
-        },
-        "weight": 4096
-      },
-      {
-        "filter": {
-          "bool": {
-            "must": [
-              {
-                "match": {
-                  "last_name": {
-                    "query": 'last name',
-                    "_name": '4_dim_lst'
-                  }
-                }
-              }
-            ]
-          }
-        },
-        "weight": 2048
-      },
-      {
-        "filter": {
-          "bool": {
-            "must": [
-              {
-                "match": {
-                  "last_name": {
-                    "query": 'last name',
-                    "_name": '5_pho_lst'
-                  }
-                }
-              }
-            ]
-          }
-        },
-        "weight": 1024
-      },
-      {
-        "filter": {
-          "bool": {
-            "must": [
-              {
-                "match": {
-                  "last_name": {
-                    "query": 'last name',
-                    "_name": '6_fz_lst'
-                  }
-                }
-              }
-            ]
-          }
-        },
-        "weight": 512
-      },
-      {
-        "filter": {
-          "bool": {
-            "must": [
-              {
-                "match": {
-                  "last_name": {
-                    "query": 'last name',
-                    "_name": '7_prt_lst'
-                  }
-                }
-              }
-            ]
-          }
-        },
-        "weight": 256
-      },
-      {
-        "filter": {
-          "multi_match": {
-            "query": 'last name',
-            "operator": 'and',
-            "fields": %w[
-              first_name
-              last_name
-            ],
-            "fuzziness": '2',
-            "_name": '7_mlt_fz'
-          }
-        },
-        "weight": 200
-      }
-    ]
-  end
-
-  def first_name_functions_part_one
-    [
-      {
-        "filter": {
-          "bool": {
-            "must": [
-              {
-                "match": {
-                  "first_name": {
-                    "query": 'first name',
-                    "_name": '1_m_fst'
-                  }
-                }
-              }
-            ]
-          }
-        },
-        "weight": 16_384
-      },
-      {
-        "filter": {
-          "bool": {
-            "must": [
-              {
-                "multi_match": {
-                  "query": 'suffix',
-                  "operator": 'and',
-                  "fields": %w[
-                    last_name
-                    suffix
-                  ],
-                  "type": 'cross_fields',
-                  "_name": '2_mlt_last_suffix'
-                }
-              }
-            ]
-          }
-        },
-        "weight": 8192
-      },
-      {
-        "filter": {
-          "multi_match": {
-            "query": 'first name',
-            "operator": 'and',
-            "fields": [
-              'akas.first_name',
-              'akas.last_name'
-            ],
-            "type": 'cross_fields',
-            "_name": '3_mlt_aka'
-          }
-        },
-        "weight": 4096
-      },
-      {
-        "filter": {
-          "bool": {
-            "must": [
-              {
-                "match": {
-                  "first_name.diminutive": {
-                    "query": 'first name',
-                    "_name": '4_dim_fst'
-                  }
-                }
-              }
-            ]
-          }
-        },
-        "weight": 2048
-      },
-      {
-        "filter": {
-          "bool": {
-            "must": [
-              {
-                "match": {
-                  "first_name.phonetic": {
-                    "query": 'first name',
-                    "_name": '5_pho_fst'
-                  }
-                }
-              }
-            ]
-          }
-        },
-        "weight": 1024
-      },
-      {
-        "filter": {
-          "bool": {
-            "must": [
-              {
-                "fuzzy": {
-                  "first_name": {
-                    "value": 'first name',
-                    "fuzziness": '5',
-                    "prefix_length": '1',
-                    "max_expansions": '25',
-                    "_name": '6_fz_fst'
-                  }
-                }
-              }
-            ]
-          }
-        },
-        "weight": 512
-      },
-      {
-        "filter": {
-          "bool": {
-            "must": [
-              {
-                "match": {
-                  "first_name_ngram": {
-                    "query": 'first name',
-                    "minimum_should_match": '25%',
-                    "_name": '7_prt_fst'
-                  }
-                }
-              }
-            ]
-          }
-        },
-        "weight": 256
-      },
-      {
-        "filter": {
-          "multi_match": {
-            "query": 'first name',
-            "operator": 'and',
-            "fields": %w[
-              first_name
-              last_name
-            ],
-            "fuzziness": '2',
-            "_name": '7_mlt_fz'
-          }
-        },
-        "weight": 200
-      }
-    ]
-  end
-
-  def full_name_functions_part_two
-    [
-      {
-        "filter": {
-          "bool": {
-            "must": [
-              {
-                "match": {
-                  "last_name": {
-                    "query": 'first name',
-                    "_name": '9a_rev_lst'
-                  }
-                }
-              },
-              {
-                "match": {
-                  "first_name": {
-                    "query": 'last name',
-                    "_name": '9a_rev_fst'
-                  }
-                }
-              }
-            ]
-          }
-        },
-        "weight": 128
-      },
-      {
-        "filter": {
-          "bool": {
-            "must": [
-              {
-                "match": {
-                  "last_name": {
-                    "query": 'first name',
-                    "_name": '9b_rev_prt_lst'
-                  }
-                }
-              },
-              {
-                "match": {
-                  "first_name_ngram": {
-                    "query": 'last name',
-                    "minimum_should_match": '25%',
-                    "_name": '9b_rev_prt_fst'
-                  }
-                }
-              }
-            ]
-          }
-        },
-        "weight": 64
-      },
-      {
-        "filter": {
-          "bool": {
-            "must": [
-              {
-                "match": {
-                  "last_name": {
-                    "query": 'last name',
-                    "_name": '10_dup_lst'
-                  }
-                }
-              },
-              {
-                "match": {
-                  "first_name": {
-                    "query": 'last name',
-                    "_name": '10_dup_fst'
-                  }
-                }
-              }
-            ]
-          }
-        },
-        "weight": 32
-      },
-      {
-        "filter": {
-          "bool": {
-            "must": [
-              {
-                "match": {
-                  "last_name_ngram": {
-                    "query": 'first name',
-                    "minimum_should_match": '25%',
-                    "_name": '11_rev_prt_lst'
-                  }
-                }
-              },
-              {
-                "match": {
-                  "first_name_ngram": {
-                    "query": 'last name',
-                    "minimum_should_match": '25%',
-                    "_name": '11_rev_prt_fst'
-                  }
-                }
-              }
-            ]
-          }
-        },
-        "weight": 16
-      },
-      {
-        "filter": {
-          "match": {
-            "last_name": {
-              "query": 'last name',
-              "_name": '8_m_lst'
-            }
-          }
-        },
-        "weight": 12
-      },
-      {
-        "filter": {
-          "match": {
-            "first_name": {
-              "query": 'first name',
-              "_name": 'xb_m_fst'
-            }
-          }
-        },
-        "weight": 8
-      }
-    ]
-  end
-
-  def last_name_functions_part_two
-    [
-      {
-        "filter": {
-          "bool": {
-            "must": [
-              {
-                "match": {
-                  "first_name": {
-                    "query": 'last name',
-                    "_name": '9a_rev_fst'
-                  }
-                }
-              }
-            ]
-          }
-        },
-        "weight": 128
-      },
-      {
-        "filter": {
-          "bool": {
-            "must": [
-              {
-                "match": {
-                  "first_name_ngram": {
-                    "query": 'last name',
-                    "minimum_should_match": '25%',
-                    "_name": '9b_rev_prt_fst'
-                  }
-                }
-              }
-            ]
-          }
-        },
-        "weight": 64
-      },
-      {
-        "filter": {
-          "bool": {
-            "must": [
-              {
-                "match": {
-                  "last_name": {
-                    "query": 'last name',
-                    "_name": '10_dup_lst'
-                  }
-                }
-              },
-              {
-                "match": {
-                  "first_name": {
-                    "query": 'last name',
-                    "_name": '10_dup_fst'
-                  }
-                }
-              }
-            ]
-          }
-        },
-        "weight": 32
-      },
-      {
-        "filter": {
-          "bool": {
-            "must": [
-              {
-                "match": {
-                  "first_name_ngram": {
-                    "query": 'last name',
-                    "minimum_should_match": '25%',
-                    "_name": '11_rev_prt_fst'
-                  }
-                }
-              }
-            ]
-          }
-        },
-        "weight": 16
-      },
-      {
-        "filter": {
-          "match": {
-            "last_name": {
-              "query": 'last name',
-              "_name": '8_m_lst'
-            }
-          }
-        },
-        "weight": 12
-      }
-    ]
-  end
-
-  def first_name_functions_part_two
-    [
-      {
-        "filter": {
-          "bool": {
-            "must": [
-              {
-                "match": {
-                  "last_name": {
-                    "query": 'first name',
-                    "_name": '9a_rev_lst'
-                  }
-                }
-              }
-            ]
-          }
-        },
-        "weight": 128
-      },
-      {
-        "filter": {
-          "bool": {
-            "must": [
-              {
-                "match": {
-                  "last_name": {
-                    "query": 'first name',
-                    "_name": '9b_rev_prt_lst'
-                  }
-                }
-              }
-            ]
-          }
-        },
-        "weight": 64
-      },
-      {
-        "filter": {
-          "bool": {
-            "must": [
-              {
-                "match": {
-                  "last_name_ngram": {
-                    "query": 'first name',
-                    "minimum_should_match": '25%',
-                    "_name": '11_rev_prt_lst'
-                  }
-                }
-              }
-            ]
-          }
-        },
-        "weight": 16
-      },
-      {
-        "filter": {
-          "match": {
-            "first_name": {
-              "query": 'first name',
-              "_name": 'xb_m_fst'
-            }
-          }
-        },
-        "weight": 8
-      }
-    ]
-  end
-
-  def fs_last_name_query
-    {
-      "size": '10',
-      "track_scores": 'true',
-      "sort": [
-        {
-          "_score": 'desc',
-          "last_name": 'asc',
-          "first_name": 'asc',
-          "_uid": 'desc'
-        }
-      ],
-      "min_score": '2.5',
-      "query": {
-        "function_score": {
-          "query": {
-            "bool": {
-              "must": [
-                {
-                  "match": {
-                    "legacy_descriptor.legacy_table_name": {
-                      "query": 'CLIENT_T',
-                      "_name": 'q_cli'
-                    }
-                  }
-                }
-              ]
-            }
-          },
-          "functions": last_name_functions,
-          "score_mode": 'sum',
-          "boost_mode": 'sum'
-        }
-      },
-      "_source": source,
-      "highlight": highlight
-    }.as_json
-  end
-
-  def fs_full_name_query_part_one
-    {
-      "size": '10',
-      "track_scores": 'true',
-      "sort": [
-        {
-          "_score": 'desc',
-          "last_name": 'asc',
-          "first_name": 'asc',
-          "_uid": 'desc'
-        }
-      ],
-      "min_score": '2.5',
-      "query": {
-        "function_score": {
-          "query": {
-            "bool": {
-              "must": [
-                {
-                  "match": {
-                    "legacy_descriptor.legacy_table_name": {
-                      "query": 'CLIENT_T',
-                      "_name": 'q_cli'
-                    }
-                  }
-                }
-              ]
-            }
-          },
-          "functions": full_name_functions_part_one,
-          "score_mode": 'sum',
-          "boost_mode": 'sum'
-        }
-      },
-      "_source": source,
-      "highlight": highlight
-    }.as_json
-  end
-
-  def fs_full_name_without_suffix_query_part_one
-    {
-      "size": '10',
-      "track_scores": 'true',
-      "sort": [
-        {
-          "_score": 'desc',
-          "last_name": 'asc',
-          "first_name": 'asc',
-          "_uid": 'desc'
-        }
-      ],
-      "min_score": '2.5',
-      "query": {
-        "function_score": {
-          "query": {
-            "bool": {
-              "must": [
-                {
-                  "match": {
-                    "legacy_descriptor.legacy_table_name": {
-                      "query": 'CLIENT_T',
-                      "_name": 'q_cli'
-                    }
-                  }
-                }
-              ]
-            }
-          },
-          "functions": full_name_without_suffix_functions_part_one,
-          "score_mode": 'sum',
-          "boost_mode": 'sum'
-        }
-      },
-      "_source": source,
-      "highlight": highlight
-    }.as_json
-  end
-
-  def fs_last_name_query_part_one
-    {
-      "size": '10',
-      "track_scores": 'true',
-      "sort": [
-        {
-          "_score": 'desc',
-          "last_name": 'asc',
-          "first_name": 'asc',
-          "_uid": 'desc'
-        }
-      ],
-      "min_score": '2.5',
-      "query": {
-        "function_score": {
-          "query": {
-            "bool": {
-              "must": [
-                {
-                  "match": {
-                    "legacy_descriptor.legacy_table_name": {
-                      "query": 'CLIENT_T',
-                      "_name": 'q_cli'
-                    }
-                  }
-                }
-              ]
-            }
-          },
-          "functions": last_name_functions_part_one,
-          "score_mode": 'sum',
-          "boost_mode": 'sum'
-        }
-      },
-      "_source": source,
-      "highlight": highlight
-    }.as_json
-  end
-
-  def fs_first_name_query_part_one
-    {
-      "size": '10',
-      "track_scores": 'true',
-      "sort": [
-        {
-          "_score": 'desc',
-          "last_name": 'asc',
-          "first_name": 'asc',
-          "_uid": 'desc'
-        }
-      ],
-      "min_score": '2.5',
-      "query": {
-        "function_score": {
-          "query": {
-            "bool": {
-              "must": [
-                {
-                  "match": {
-                    "legacy_descriptor.legacy_table_name": {
-                      "query": 'CLIENT_T',
-                      "_name": 'q_cli'
-                    }
-                  }
-                }
-              ]
-            }
-          },
-          "functions": first_name_functions_part_one,
-          "score_mode": 'sum',
-          "boost_mode": 'sum'
-        }
-      },
-      "_source": source,
-      "highlight": highlight
-    }.as_json
-  end
-
-  def fs_full_name_query_part_two
-    {
-      "size": '10',
-      "track_scores": 'true',
-      "sort": [
-        {
-          "_score": 'desc',
-          "last_name": 'asc',
-          "first_name": 'asc',
-          "_uid": 'desc'
-        }
-      ],
-      "min_score": '2.5',
-      "query": {
-        "function_score": {
-          "functions": full_name_functions_part_two,
-          "score_mode": 'sum',
-          "boost_mode": 'sum'
-        }
-      },
-      "_source": source,
-      "highlight": highlight
-    }.as_json
-  end
-
-  def fs_last_name_query_part_two
-    {
-      "size": '10',
-      "track_scores": 'true',
-      "sort": [
-        {
-          "_score": 'desc',
-          "last_name": 'asc',
-          "first_name": 'asc',
-          "_uid": 'desc'
-        }
-      ],
-      "min_score": '2.5',
-      "query": {
-        "function_score": {
-          "functions": last_name_functions_part_two,
-          "score_mode": 'sum',
-          "boost_mode": 'sum'
-        }
-      },
-      "_source": source,
-      "highlight": highlight
-    }.as_json
-  end
-
-  def fs_first_name_query_part_two
-    {
-      "size": '10',
-      "track_scores": 'true',
-      "sort": [
-        {
-          "_score": 'desc',
-          "last_name": 'asc',
-          "first_name": 'asc',
-          "_uid": 'desc'
-        }
-      ],
-      "min_score": '2.5',
-      "query": {
-        "function_score": {
-          "functions": first_name_functions_part_two,
-          "score_mode": 'sum',
-          "boost_mode": 'sum'
-        }
-      },
-      "_source": source,
-      "highlight": highlight
-    }.as_json
-  end
-
   def fs_full_name_query
     {
       "size": '10',
@@ -1816,61 +498,6 @@ module PersonSearchByNameQueryBuilderHelper
     }.as_json
   end
 
-  def fs_full_name_without_suffix_approx_age_months_gender_query
-    {
-      "size": '10',
-      "track_scores": 'true',
-      "sort": [
-        {
-          "_score": 'desc',
-          "last_name": 'asc',
-          "first_name": 'asc',
-          "_uid": 'desc'
-        }
-      ],
-      "min_score": '2.5',
-      "query": {
-        "function_score": {
-          "query": {
-            "bool": {
-              "must": [
-                {
-                  "match": {
-                    "legacy_descriptor.legacy_table_name": {
-                      "query": 'CLIENT_T',
-                      "_name": 'q_cli'
-                    }
-                  }
-                },
-                {
-                  "range": {
-                    "date_of_birth": {
-                      "gte": (Date.current - 12.months - 6.months).iso8601,
-                      "lte": (Date.current - 12.months + 6.months).iso8601,
-                      "format": 'yyyy-MM-dd'
-                    }
-                  }
-                },
-                {
-                  "query_string": {
-                    "default_field": 'gender',
-                    "query": 'female',
-                    "boost": '1'
-                  }
-                }
-              ]
-            }
-          },
-          "functions": full_name_without_suffix_functions,
-          "score_mode": 'sum',
-          "boost_mode": 'sum'
-        }
-      },
-      "_source": source,
-      "highlight": highlight
-    }.as_json
-  end
-
   def fs_full_name_approx_age_years_gender_query
     {
       "size": '10',
@@ -1917,6 +544,61 @@ module PersonSearchByNameQueryBuilderHelper
             }
           },
           "functions": full_name_functions,
+          "score_mode": 'sum',
+          "boost_mode": 'sum'
+        }
+      },
+      "_source": source,
+      "highlight": highlight
+    }.as_json
+  end
+
+  def fs_full_name_without_suffix_approx_age_months_gender_query
+    {
+      "size": '10',
+      "track_scores": 'true',
+      "sort": [
+        {
+          "_score": 'desc',
+          "last_name": 'asc',
+          "first_name": 'asc',
+          "_uid": 'desc'
+        }
+      ],
+      "min_score": '2.5',
+      "query": {
+        "function_score": {
+          "query": {
+            "bool": {
+              "must": [
+                {
+                  "match": {
+                    "legacy_descriptor.legacy_table_name": {
+                      "query": 'CLIENT_T',
+                      "_name": 'q_cli'
+                    }
+                  }
+                },
+                {
+                  "range": {
+                    "date_of_birth": {
+                      "gte": (Date.current - 12.months - 6.months).iso8601,
+                      "lte": (Date.current - 12.months + 6.months).iso8601,
+                      "format": 'yyyy-MM-dd'
+                    }
+                  }
+                },
+                {
+                  "query_string": {
+                    "default_field": 'gender',
+                    "query": 'female',
+                    "boost": '1'
+                  }
+                }
+              ]
+            }
+          },
+          "functions": full_name_without_suffix_functions,
           "score_mode": 'sum',
           "boost_mode": 'sum'
         }
