+++ spec/support/person_search_result_builder.rb
@@ -5,6 +5,9 @@ require 'support/helpers/query_builder_helper'
 require 'support/helpers/person_search_query_builder_helper'
 require 'support/helpers/person_search_ssn_query_builder_helper'
 require 'support/helpers/person_search_by_name_query_builder_helper'
+require 'support/helpers/person_search_by_last_name_query_builder_helper'
+require 'support/helpers/person_search_by_name_query_builder_part_one_helper'
+require 'support/helpers/person_search_by_name_query_builder_part_two_helper'
 require 'support/helpers/person_search_by_date_of_birth_query_builder_helper'
 require 'support/helpers/person_search_by_approximate_age_query_builder_helper'
 require 'support/helpers/person_search_gender_query_builder_helper'
@@ -14,6 +17,9 @@ class PersonSearchResultBuilder
   include PersonSearchQueryBuilderHelper
   include PersonSearchSsnQueryBuilderHelper
   include PersonSearchByNameQueryBuilderHelper
+  include PersonSearchByLastNameQueryBuilderHelper
+  include PersonSearchByNameQueryBuilderPartOneHelper
+  include PersonSearchByNameQueryBuilderPartTwoHelper
   include PersonSearchByDateOfBirthQueryBuilderHelper
   include PersonSearchByApproximateAgeQueryBuilderHelper
   include PersonSearchGenderQueryBuilderHelper
